samples/market-research-agent/requirements.txt
# Core LLM Libraries
openai>=1.12.0
anthropic>=0.18.0

# Web Scraping & Data Gathering
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Search API
tavily-python>=0.3.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
